requirements.txt
flask==3.0.0
requests==2.32.3
flask-mail
python-dotenv